Move Player entity into PlayerAggregate
