+++ src/TournamentManagement/TournamentManagement.Domain/PlayerAggregate/Player.cs
@@ -1,7 +1,7 @@
 ﻿using DomainDesign.Common;
 using System;
 
-namespace TournamentManagement.Domain
+namespace TournamentManagement.Domain.PlayerAggregate
 {
 	public class Player : Entity<PlayerId>
 	{
